Fix include directives for system headers
When a system header is included, it will be found through the "/builtins/"
prefix, as this is how smokegen bundles the clang headers with the
executable. If a symbol from one of these files is used, then the generated
x##.cpp file will have "#include </builtins/....h>" in it, which is
incorrect. With this change, the leading "/builtins/" is stripped.